Move `ActionAfterDisconnect` to Talpid package
